backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
openai>=1.40.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
reportlab>=4.0.0
pydantic>=2.0.0
qrcode[pil]>=7.4
Pillow>=10.0
requests>=2.31
httpx>=0.27.0
striprtf>=0.0.26
beautifulsoup4>=4.12
pytesseract>=0.3.10
chainlit>=1.0.0
